abstracts fastly from job
