+++ worker/jobTypes/publishDochubJob.js
@@ -2,7 +2,7 @@ const workerUtils = require('../utils/utils');
 const validator = require('validator');
 
 const invalidJobDef = new Error('job not valid');
-const mongo = require('../utils/mongo');
+const FastlyJob = require('../jobTypes/fastlyJob').FastlyJobClass
 
 //anything that is passed to an exec must be validated or sanitized
 //we use the term sanitize here lightly -- in this instance this // ////validates
@@ -37,6 +37,17 @@ function safePublishDochub(currentJob) {
   throw invalidJobDef;
 }
 
+async function startFastly(job) {
+  // retrieve Fastly service
+  var fastly = require('fastly')(`${process.env.FASTLY_TOKEN}`)
+
+  // connect to MongoDB dochub database
+  const MongoClient = require("mongodb").MongoClient;
+  assert = require("assert")
+
+  job.connectAndUpsert(MongoClient, fastly);
+}
+
 async function runPublishDochub(currentJob) {
 
   workerUtils.logInMongo(currentJob, ' ** Running dochub-fastly migration');
@@ -52,55 +63,9 @@ async function runPublishDochub(currentJob) {
     throw invalidJobDef;
   }
 
-  // retrieve Fastly service
-  var fastly = require('fastly')(`${process.env.FASTLY_TOKEN}`)
+  const job = new FastlyJob(currentJob);
 
-  var XMLHttpRequest = require("xmlhttprequest").XMLHttpRequest;
-  var req = new XMLHttpRequest();
-
-  // connect to MongoDB dochub database
-  const MongoClient = require("mongodb").MongoClient;
-  assert = require("assert")
-
-  const url = `mongodb+srv://${process.env.MONGO_ATLAS_USERNAME}:${process.env.MONGO_ATLAS_PASSWORD}@cluster0-ylwlz.mongodb.net/dochub?authSource=admin`;
-  
-  MongoClient.connect(mongo.url, function(err, client) {
-    assert.equal(null, err);
-
-    const db = client.db("dochub");
-
-    var cursor = db.collection('keys').find({});
-
-    function iterateFunc(doc) {
-      const page = "https://dochub.mongodb.org/core/" + doc.name;
-      var request = require('request');
-      var r = request.get(page, function(err, res, body) {
-        if (res != null) {
-          if (res.status != 404) {
-            const options = {
-              item_value: doc.url
-            }
-
-            fastly.request('PUT', '/service/0U4FLNfta0jDgmrSFA193k/dictionary/2FoAatLRziZlxb6aTwnRWs/item/'+doc.name, options, function (err, obj) {
-              if (err) return console.dir(err);
-              console.dir(obj);
-            });
-          } else {
-            console.log("Bad URL: ", page, res.status);
-          }
-        }
-      });
-    }
-
-    function errorFunc(error) {
-      console.log(error);
-    }
-
-    cursor.forEach(iterateFunc, errorFunc);
-
-
-    client.close();
-  });
+  await startFastly(job);
 }
 
 module.exports = {
