+++ worker/tests/mongo/globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoDBName":"jest","mongoUri":"mongodb://127.0.0.1:49246/jest"}
+{"mongoDBName":"jest","mongoUri":"mongodb://127.0.0.1:56854/jest"}
